Add files via upload
Image Classification is the process of passing an image to a machine learning algorithm into a vector form and the algorithm will identify the image by matching it with the given sample or label the image. 

The image is converted into a vector of numbers ( pixel of the image) and given to the algorithm. I have used the KNN for classification. but we can also use another classification algorithm  as well as neural network.